Added mode to automatically play the game
Added a simulation mode that automatically plays the game.  Have a dumb
mode that just guesses randomly and a smarter mode that hunts as soon as
it finds a 'hit' until it destroys something.
Added ability to generate board state upon completion so I can generate
a bunch of games sorted by the number of moves it takes.